+++ guillotina/api/ws.py
@@ -20,8 +20,23 @@ import asyncio
 import ujson
 
 
-@configure.service(context=IContainer, method='GET', permission='guillotina.AccessContent',
-                   name='@wstoken')
+@configure.service(
+    context=IContainer, method='GET',
+    permission='guillotina.AccessContent', name='@wstoken',
+    summary='Return a web socket token',
+    responses={
+        "200": {
+            "description": "The new token",
+            "schema": {
+                "properties": {
+                    "token": {
+                        "type": "string",
+                        "required": True
+                    }
+                }
+            }
+        }
+    })
 class WebsocketGetToken(Service):
     _websockets_ttl = 60
 
@@ -54,8 +69,10 @@ class WebsocketGetToken(Service):
         }
 
 
-@configure.service(context=IContainer, method='GET', permission='guillotina.AccessContent',
-                   name='@ws')
+@configure.service(
+    context=IContainer, method='GET',
+    permission='guillotina.AccessContent', name='@ws',
+    summary='Make a web socket connection')
 class WebsocketsView(Service):
 
     async def __call__(self):